+++ Wiki/ReleaseNotes.md
@@ -6,6 +6,7 @@
 - Added batch capability to IMessageTracker.
 - Added batch capabilities upsert to IAtomicCollectionWrapper.
 - Added batch capabilities to AtomicReadmodelProjectorHelper.
+- Added FindManyAndProjectAsync in atomic collection reader.
 
 ## 7.17.2